+++ src/GlobalTrafficTable.h
@@ -22,8 +22,8 @@ using namespace std;
 struct Communication {
   int src;			// ID of the source node (PE)
   int dst;			// ID of the destination node (PE)
-  float pir;			// Packet Injection Rate for the link
-  float por;			// Probability Of Retransmission for the link
+  double pir;			// Packet Injection Rate for the link
+  double por;			// Probability Of Retransmission for the link
   int t_on;			// Time (in cycles) at which activity begins
   int t_off;			// Time (in cycles) at which activity ends
   int t_period;		        // Period after which activity starts again
